pyproject.toml
[project]
name = "chesstrm"
version = "0.1.0"
description = "Recursive Transformer Model for Chess"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch",
    "python-chess",
    "h5py",
    "numpy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ruff",
    "pytest",
]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
